add config & helpers
