+++ src/js/model.js
@@ -1,17 +1,14 @@
 import { async } from 'regenerator-runtime';
-
+import { API_URL } from './config.js';
+import { getJSON } from './helpers.js';
 export const state = {
   recipe: {},
 };
 
 export const loadRecipe = async function (id) {
   try {
-    const res = await fetch(
-      `https://forkify-api.herokuapp.com/api/v2/recipes/${id}`
-    );
-    const data = await res.json();
+    const data = await getJSON(`${API_URL}/${id}`)
     let { recipe } = data.data;
-    if (!res.ok) throw new Error(`${data.message} (${res.status})`);
     state.recipe = {
       cooking_time: recipe.cooking_time,
       id: recipe.id,
@@ -24,6 +21,6 @@ export const loadRecipe = async function (id) {
     };
     console.log(state.recipe);
   } catch (error) {
-    alert(error);
+    console.log(error)
   }
 };
