+++ src/js/views/recipeView.js
@@ -7,7 +7,7 @@ class RecipeView {
   render(data) {
     this.#data = data;
     const markup = this.#generateMarkup();
-    this.#clear;
+    this.#clear();
     this.#parentElement.insertAdjacentHTML('afterbegin', markup);
   }
